Fix image download issues

+++ include/progress_bar.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 #include <string>
 
 #include <indicators/progress_bar.hpp>
@@ -11,7 +11,7 @@ namespace kepub {
 
 class KEPUB_EXPORT ProgressBar {
  public:
-  explicit ProgressBar(std::int32_t max_progress,
+  explicit ProgressBar(std::size_t max_progress,
                        const std::string &postfix_text);
 
   void set_postfix_text(const std::string &postfix_text);
